Extract ValueRenderer to standalone file and add displayMode prop
Move ValueRenderer component and resolve() helper from ResultDisplay.tsx
into ValueRenderer.tsx. Add DisplayMode type ('inline' | 'expanded' |
'auto' | 'inline-hint') to result-renderers.ts and thread it through
the renderer contract. No behavioral changes — pure refactor.
